Allow client to request only current role appointments
The prime minister page does not need the former role appointments to
render, so there is no need to retrieve them.

By adding an argument to the `role_appointments` link field, we can
reduce the number of documents whose links need to be further expanded.

An index for editions with a current value in the details was added in
e1b1955eb4e9e72b49674f72f690bb355c00f1c2.