Don't require a rendering app for SpecialRoutePublisher
Requiring a rendering app prevents special route publisher from
publishing routes that don't have one such as gone and redirects.

My motivation for making this change is so that we can use
SpecialRoutePublisher to publish a gone route for /chat.

This will make the Rails logger line a little bit janky with the
"routing to ". But this seems such a modest issue that it isn't worth
writing extra code.