Add special route for GOV.UK Chat
GOV.UK Chat was an application hosted at www.gov.uk/chat. Its routing
was configured at the CDN level so it did not have a presence in
Publishing API.

As part of retiring the application we want the URLs to return a 410
status code. The most idiomatic way to do this for GOV.UK Publishing
seems to be to create a content item for it.

GOV.UK Chat had a number of routes used but I didn't want to set a
prefix route for it as that'd mean anything under /chat would get a 410.
Instead I opted for just the homepage and the content pages with the
expectation that isn't any issue for application routes to return 404.

Note: title and description are required to be null for a gone schema,
hence the explicit declarations [1]

[1]: https://github.com/alphagov/publishing-api/blob/61eb70f361de537da3cdab41eeb810be25c4b5f9/content_schemas/formats/gone.jsonnet#L6-L7