Only parse details fields that are requested
At the moment, we are putting the entire contents of the `details` field
through the `DetailsPresenter`, which is sub-optimal as we are parsing
values that clients are not requesting.

By using lookaheads, we can determine which fields from `details` have
been requested and only parse those.

In a previous prototype, we had only transformed the govspeak in the
`body` field. However the schemas can permit mixed govspeak/HTML content
in any details field. We therefore need to parse all items within
`details`, as is already done using the `DetailsPresenter`.

This also allows removal of the optimisations for `change_history` added
in 2caf00ec9770437c66298f736a55713dd15c6c59 since we can now filter the
details to only include that field when requested.