Do not loop through details unless embed links exist
At the moment, we are looping through each item in `details` to look for
any embedded content references. This is sub-optimal as we are searching
through content for a tag that we know won't be there, if there is
nothing to embed in the document.

We should skip doing this if there is nothing that could possibly be
embedded, which we know by looking in the document's links.