Reduce code execution in `DetailsPresenter`
This presenter currently performs some of the same code in different
methods that determine the type of content in the field, and looks for
content types even when we know there won't be any present.

Therefore refactoring this presenter to perform as little code execution
as possible to order to determine the correct outcome for the content
type given.

This performance improvement is particularly important for GraphQL, as
we will be converting govspeak to HTML at render-time.