Remove redundant test
There is no schema that permits this format for specifying the content
is govspeak (all schemas require this in an array), so the test can be
removed.

This behaviour is confired by ADR-003 [1] and the definition of
`multiple_content_types` in the shared Publishing API base schema
definition.

Removing this test (and support for multi-part content to not be
included in an array)  allows us to simplify the `DetailsPresenter`, as
we will only ever be dealing with strings or arrays of hashes.

Also updating a validator to ensure no content with this incorrect
syntax can be published in the future.

1: https://github.com/alphagov/publishing-api/blob/main/docs/arch/adr-003-representation-for-multiple-content-types.md