Only SELECT the root Edition's columns we need
We're hoping to speed up our GraphQL responses by reducing the amount of
data retrieved from the database. By design we already only send the
client the data they've requested, but until now that meant that we've
been fetching a lot of data from the database that we don't actually
make use of in serving the response to the client.

SELECTing a subset of the root Edition's columns has an impact on how we
can make use of the Edition instance. Until now we've been freely
lazy-loading a couple of the Edition's associations as and when they've
been needed (its document and its unpublishing), but that becomes more
complicated now (for one thing, we need to be sure to SELECT the foreign
keys needed to even find those associations). Our first go at responding
to this problem is to determine ahead of time, using the GraphQL
`lookahead`, which associations' columns are needed and to SELECT them
from their respective JOINed tables. Active Record, for better or for
worse, dynamically assigns those columns as attributes on the Edition
instance (so it seems like a decent idea to `AS`-alias the more
generic-named columns that aren't native to Edition itself).

A note of the changes to the withdrawn_notice resolver: we believe that
the withdrawn_notice is only really supported for the root edition and
not for editions in general (e.g. linked-to editions). That's the reason
we've taken the liberty to make the changes we have. We don't expect to
JOIN unpublishings to the other editions handled later on in serving the
GraphQL query, so the resolver's new `respond_to?` early exit should
prevent the whole query from raising an error if a client was to
mistakenly request a linked-to edition's withdrawn_notice.