Only SELECT the root Edition's columns we need
We're hoping to speed up our GraphQL responses by reducing the amount of
data retrieved from the database. By design we already only send the
client the data they've requested, but until now that meant that we've
been fetching a lot of data from the database that we don't actually
make use of in serving the response to the client.