Allow Edition to provide content_id, locale itself
I'm not 100% sure about this change.

We're experimenting with SELECTing subsets of an edition's columns and
JOINed-tables' columns to improve the performance of our GraphQL
resolvers. When you SELECT JOIN columns, Active Record helpfully just
seems to dynamically add them as attributes to the model you're working
with, which means that, e.g.

  SELECT editions.*, documents.content_id, documents.locale FROM editions
  INNER JOIN documents ON document.id = editions.document_id

would result in instances of Edition with their own, native #content_id
and #locale attributes. In those specific examples, the existing
`delegate` doesn't appear to acknowledge these attributes and still
attempts to lazy-load the edition's document to fetch them. Since that
lazy-loading kinda defeats the point of the approach we're experimenting
with, here's an alternative to `delegate` that first checks for
pre-existing attributes of the same name.

There could be better approaches than this worth exploring in the long
run if we do stick with this SELECTing optimisation. (I don't think it's
wise to use these Active Record models in such dramatically different
ways in different parts of the codebase.)