SELECT "documents"."locale" for root Edition
This wasn't needed when we were testing against the ministers index
page, but once we tried the prime minister page `locale` was found to be
missing.

I still don't know which line of code was trying to access
Edition#locale (because the stacktrace didn't seem to say), but
SELECTing the locale made the error go away.