Rename variable to be shorter and sweeter
I originally named this variable that way because I was having trouble
remembering what a "link type" even is and what its significance might
be. We've been knee-deep in these Publishing API concepts for a couple
of months now and as a result both of those things have become second
nature and this variable name has become cumbersome.

(I'm not sure whether the naming for reverse_links_field was ever quite
accurate.)