Schema change to nested facet structure
Finder frontend and Government frontend view the nested facets as 2 separate facets. We make it easier by sending them as separate facets with a reference back to their main facet value.
Main facet label is also required to populate the subfacet label in the format `Main facet label - Sub Facet Label` for when JS is disabled.

The changes made in SP now make the use of `sub_facets` field obsolete, but the removal of that field will be done in a separate PR.