Introduce GraphqlSelections
We're working towards SELECTing a subset of an Edition's columns, plus
a subset of some JOINed associations' columns, too. Working with the
sets, arrays and hashes required to build a complex SELECT argument is
getting increasingly gross, so this object is intended to provide
a smaller and neater interface for doing that.

This also provides a constructor that converts the GraphQL query's
edition selections to database column names. We include a list of valid
database columns and a hash which maps GraphQL fields not in the list to
the database columns they rely on

Co-authored-by: Ynda Jas <yndajas@gmail.com>