SELECT subset of columns for role_appointments
This "optimisation" is specific to the ministers index type.

We've followed the patterns established in the root Edition and the more
generic Link sources. The big differences are mostly down to
* the larger number of JOINs used in this query
* the convenience of only needing to handle link set links here
* this data loader fetches a person's roles but is actually invoked from
  the person's role_appointments field... weird

Because of the large number of JOINS, we've resorted to supplying SQL
strings in order to provide our own, hopefully clearer, join table
aliases (as opposed to leaving that to Active Record to generate).