Only SELECT person's roles columns that we need
The idea with this change is to reduce the amount of data retrieved from
the database in an attempt to speed up the overall response time to the
GraphQL client.

So far we've implemented this optimisation for the root edition,
linked-to editions and reverse-linked editions. Here we're introducing
the change for the person's role_appointments/roles fields specific to
the ministers index type.

We've followed the patterns established in the previous implementations
with the big differences being mostly:
* the larger number of JOINs needed in this query because we're
  traversing 2 levels of links
* the convenience of only needing to handle link set links here because
  we know that those are the only types of links involved in linking the
  ministers index's person-role_appointments-roles

Because of the large number of JOINS, we've resorted to supplying SQL
strings in order to provide our own, hopefully clearer, join table
aliases as opposed to leaving that to Active Record to generate.