Add document type labels to publication schema
The research and statistics finder and the official document finder render a human friendly version of the document type in the search result metadata.

Now that we are seeking to index all Whitehall documents via Publishing API, we need to make sure that we send these labels as part of the edition payload.