Revert "Do not loop through details unless embed links exist"
This reverts commit 0c78953ead3478ff19aaba3d7e1e6e32f094f8d0.

We previously introduced an optimisation to reduce the amount of text
parsing needed for embedded content links, however this resulted in a
database query instead of the text parsing.

For some pages (e.g. the Prime Minister page), this resulted in GraphQL
reequests making 176 queries instead of 15, which was slower than
parsing the text.

Reverting the change, to restore the previous behaviour.