Reduce query counts in GraphQL dataloader sources
We found that the previous eager loading approach wasn't using joins
effectively, so we're more manually crafting the queries to join in an
efficient way. We also found that some associations we're still lazy
loaded after the initial queries - this new approach avoids that issue

For now we're selecting all columns from the editions table. We'll
reduce the selections down to a subset in a later change

The change to using a UNION affects the order of the resulting editions,
so we've added a test to document our expectations

Co-authored-by: Mike Patrick <mike.patrick2@digital.cabinet-office.gov.uk>