Add configuration for read replica
The read replica is a separate deployment of this application, which
connects to both the primary database (e.g. for writing user records
after authentication) and the replica database (for all reads).

Toggling of the application is done by changing the `RAILS_ENV`
environment variable from `production` to `production_replica`.