Query for getting embedded editions
When looking at optimising our embedded editions process,
I noticed that we are using quite a complicated set of
db queries to get the embedded editions from a host edition
- as there the GraphQL team are looking to reduce
db queries [1] I thought I would take a look. This new
query I think improves things:
1. If I've worked it out right, this makes one less call to db
2. It's much simpler and removed the logic around locale and state -
for embedded editions we only have `en` and we only ever want `published`
editions. This might change in the future, but seems like the simpler
we can keep things for now the better.
This will only affect calls where embed codes have been found, so won't
make much of a difference now, but should still be a bit of an improvement
if we ever scale to 100's of embeds!

[1] https://trello.com/c/ilVC6hCV