Write links(link_set_content_id) in patch link set
This is ADR-009 step 2 - Update the code to write to both columns.

PATCH /v2/links/:content_id is the only endpoint that creates or updates link set links, so I'm pretty confident this is the only place we need to dual-write to the old (link_set_id) and new (link_set_content_id) columns.