Populate link_set_content_id with rake task
We could do this with a single SQL query, like:

    update links
    set links.link_set_content_id = link_sets.content_id
    from link_sets
    where links.link_set_id is not null and links.link_set_id = link_sets.id;

But this would hold a ROW EXCLUSIVE lock on the links table for the
duration of the query. Locally, this takes around 15 minutes, and we'd
prevent any updates of the links table for that time. That's probably
best avoided, and as we don't need to add all the values at once it
feels more sensible to do them in batches.

This rake task finds 1,000 links at a time (the find_in_batches default)
and then sets their link_set_content_id to link_sets.content_id.

It takes an id to start from, in case we get mid way through and want to
restart.