Support empty change history in GraphQL
We haven't yet started supporting GraphQL in frontend apps for any
document types that request the change history from the details JSON.
However, the recent work to reduce database queries and use lookaheads
in GraphQL dataloaders has inadvertently reduced support for querying
the change history from an edition's details field. That change involved
eager loading associations and minimising selected database columns.
We've since found that when requesting a change history for an edition
with no change notes, the server returns a 500 error. The issue is that
the `ChangeHistoryPresenter` tries to get an edition's document to pass
in its ID to an Active Record query, but our custom `Edition` no longer
has the `document_id` field from which to lazy load its document
association

Meanwhile, when an edition already has change notes, presumably the
change history will already exist in the edition's details field, and
`ChangeHistoryPresenter#change_history` will defer to the existing
history rather than firing off a query to build a new one. We haven't
looked into how the change history gets updated when new change notes
are added (after the initial one), but we have confirmed that change
histories still work for editions with change notes

To fix the issue with change histories where there are no change notes,
because the change history is nested within the details field, we can't
simply add the nested field into `GraphqlSelections::FIELDS_TO_COLUMNS`
without adding complexity to that hash's consumers. So we instead always
select an edition's `document_id` and then use `edition.document_id` in
place of `edition.document` in the Active Record query. We also need the
edition's `user_facing_version` in the same query, so that too is added
to our selections. We've added these selections for all editions, so
change histories are queryable on links. We aren't aware of this being a
current requirement, but we felt it seemed plausible

We'd like to explore simplifying things (at least for GraphQL usage),
potentially moving away from using the `ChangeHistoryPresenter`, but for
now this is a quick fix

Co-authored-by: Bruce Bolt <bruce.bolt@digital.cabinet-office.gov.uk>
Co-authored-by: Callum Knights <callum.knights@digital.cabinet-office.gov.uk>
Co-authored-by: Jon Hallam <jonathan.hallam@digital.cabinet-office.gov.uk>
Co-authored-by: Mike Patrick <mike.patrick2@digital.cabinet-office.gov.uk>