Amend specialist document examples for new metadatablock property
Some tests on government frontend use these examples for their rendering tests.

I've made drug device alerts and countryside stewardship grants the 'true' cases for rendering the
metadata block.

Cma cases not having the property is acting as the 'false', i.e do not show the block.

Business finance is the scenario in which the 'show_metadata_block' property is true, but the document lacks
'important metadata'. This is achieved by removing the facets.