Use content_id as primary key in link_set associations
This changes the associations to join on link_set_content_id instead of joining through link_sets on link_set_id.

There are a couple of workarounds needed to ensure that link_set_id is still populated when links are created (either in patch link_set, or in the test factories). These should be removed when we're no longer using the link_set_id column.

Because content_id is now the pkey of LinkSet, we can't use LinkSet.last in the tests to get the most recent one anymore (it will get the one with the lexicographically last content_id, which is random). It's more explicit to look them up by content_id anyway.