Use content_id to associate actions with link sets
... instead of link_set_id

Unfortunately this means we have to give up the validation that an
actions can't be for both an edition and a link_set. I don't think this
is a huge concession though. It does mean if you look at an action
that's not got anything to do with a link set, you'll be able to call
.link_set on it, and get the link_set for the content_id of the edition.
That feels okay to me though.

This will mean we can later remove the link_set_id column from actions,
and eventually the id column from link_sets