Avoid unnecessary join in represent_downstream rake
This feels like a pretty wild rake task - there are 583,908 distinct
content_ids with taxon links in my database dump, and this would
represent them all.

I suspect this task is not needed anymore to be honest - it was added in
200e1b54 to fix a data issue, but I can't see why we'd need it
ordinarily.

I'll just remove it in a subsequent commit.