Remove represent_downstream:tagged_to_taxon rake task
I don't think this task is likely to be run very often (or ever?)

It was added in 200e1b54 to fix a link expansion issue, but it's quite
specific to that situation. It would also be very expensive to run -
most documents belong to at least one taxon, so it's almost as bad as
running represent_downstream:all.

Less code means less maintenance, so I think just delete it.