Use Renovate for Ruby version
Dependabot doesn't open PRs to upgrade Ruby. It would be good keep up to
date where possible, and at least keep updates on our radar, so this
sets up Renovate to open PRs to update Ruby. We also need to update the
version in the Dockerfile for minor and major bumps, so a note is added
to the PR description as a reminder for these upgrades

We extend Renovate's best practice config, then tell Renovate to always
open a PR (rather than directly merging branches) and to automerge
patches but only when status checks all pass. Like Dependabot,
"dependencies" and "ruby" labels are applied as appropriate

Renovate won't be enabled until govuk-helm-charts is updated:
https://github.com/alphagov/govuk-helm-charts/blob/d5d6bd5affe0524f05786c44187c4827db3c1b3a/charts/renovate/templates/configmap.yaml#l8

C.f.:
- https://github.com/alphagov/govuk-helm-charts/blob/main/renovate.json
- https://github.com/alphagov/govuk-infrastructure/blob/main/renovate.json
- https://github.com/dxw/renovate-config