GraphQL field type for Content API-a-like datetime
Timezones are hard and ActiveSupport timezone-handling is even harder.

What I'm aiming for here is consistency of the format of datetime
strings being served to clients from our GraphQL endpoint and Content
Store's API.

The main problem really comes down to Publishing API's locale being UTC
and Content Store's being Europe/London, though the inconsistency is
a little bit messier than that.

So, to match Content Store's datetimes, we need to convert our datetimes
into Content Store's timezone before serving them to clients.

The messier part (the bit that confused the hell out of me) is that
ActiveSupport::TimeWithZone (the type of object that we're typically
handling since it's what ActiveRecord will return database values as)
follows different datetime-to-string formatting rules depending on
whether the timezone is UTC or non-UTC.

The result is that even at those times of year (non-daylight savings)
when I expected Publishing API's UTC datetimes to agree with Content
Store's, the formatted strings being produced are still different:

* Publishing API's UTC `datetime#iso8601` #=> "2025-01-01T00:00:00Z"
* Content Store's GMT `datetime#iso8601` #=> "2025-01-01T00:00:00+00:00"

I just thought I'd make a note of that while I was here. Converting the
datetime into Content Store's timezone does fix that inconsistency too.

---

I've defined both the #coerce_input and #coerce_result methods for this
new type based on the advice in the docs:
https://graphql-ruby.org/type_definitions/scalars.html#custom-scalars

I'm not sure what the full breadth of use cases for #coerce_input is,
but since what we're building is a read-only API and since none of our
queries accept datetime-type arguments (would it be used for this?),
I don't believe an implementation for that method is actually needed.