Support Image#high_resolution_url in GraphQL
This is used in by govuk_publishing_components if present to render a
higher resolution image than the default. We'll need to query it in
frontend apps (e.g. Frontend's edition query for use in news articles)
in order to render the same image as when Content Store is the source