add initial schema for a Content Block Contact
we are developing a type of Contact for the Content
Block Manager - the shape and fields of the block
is still in flux,but we're assuming here that the
"modes" of a Contact will use the same underlying
structure as embedded objects.

here we add a Contact that can have many email
address, and/or telephone, objects.