Send Content Block update jobs to the HIGH_QUEUE
We’ve had a lot of experiences with content block updates
getting stuck behing bulk update jobs, which is causing
confusion for users when carrying out user research, as
well as causing intermittent failure with e2e tests.

Because users expect content block updates to be
instentaneous (or as close to instentaneous as possible)
they’re slightly different to other types of dependency
resolution updates (for example, listing pages).