Return an error for `vanish` unpublished editions
GraphQL expects all responses to be 200, even if there is an error,
therefore we cannot return a 410 as we do with Content Store.

Instead we will omit the edition data and append the unpublishing type
into the errors, so the frontend app can handle these.

The unpublishing type of `vanish` is being added in this commit, as they
are the simplest types of unpublishing as no further information is
needed by the frontend application.