Return an error for `gone` unpublished editions
GraphQL expects all responses to be 200, even if there is an error,
therefore we cannot return a 410 as we do with Content Store.

Instead we will omit the edition data and append the unpublishing type
into the errors, so the frontend app can handle these.

The unpublishing type of `gone` is being added in this commit. The
frontend applications require more information to render a 'gone'
response, since they also need information regarding the reason and/or
alternative URLs to present to the user.