ADR009 do not set link_set_id fkey
We're using link_set_content_id everywhere in publishing-api's codebase
now, so we no longer need to set link_set_id on newly created links.

This will be a breaking change for govsearch, as it operates on a dump
of publishing-api's database. We should make sure govsearch is updated
to use link_set_content_id before deploying this.

The link_set_id column will be removed entirely in a future commit.