Remove Renovate config
We're not quite ready to do the work to get this right yet (e.g. bring
the Ruby versions in sync with our own internal image versions, with
apps at the bleeding edge and gems at the opposite of the bleeding
edge). We might come back to this in a firebreak week or similar

Removing the config since we're disabling Renovate (via [Helm
Charts][]), and having it lying around would be confusing

[Helm Charts]: https://github.com/alphagov/govuk-helm-charts/pull/3182