Add script to diff GraphQL and Content Store
This downloads a Content Store and GraphQL HTML response for the same
page and diffs the two

We need to remove the hashes that Rails adds to various HTML tags. The
hash removal isn't fully watertight. It will:
- look within HTML tags for any `aria-labelledby`, `for`, or `id`
  properties and replace those characters with "HASH" if their value
  ends in eight lower case alphabetical or numeric characters. It's
  possible that it would transform a non-hashed value, but the
  likelihood of this replacement removing a meaningful diff between
  Content Store- and GraphQL-sourced renders feels very low
- remove hashed values from nonce properties in script tags

It also removes any instances of `?graphql=true`, which we don't care
about, and the updated-at meta tag content, since this will differ
between Content Store and Publishing API, but that's not a difference we
care about

A few functions are written in a separate file. These will be reused in
a separate script for diffing multiple pages in a single run