GraphQL Reverse link from role to organisations
We can use the existing link structure to lookup the organisations that
a role belongs to. This relationship is represented by a forward link
from an organisation via a "ordered_roles" link to a number of roles,
but we can follow it in reverse to go back from a role to a number of
organisations it belongs to.

Combining this with adding content_ids to the organisations in the
ministers index query lets us check whether a role belongs to a
ministerial department in a much more efficient way than querying all of
the roles in every department.