Fix sed command
When using `-i` on macOS, the next argument is taken as a suffix to add
to a backup copy before the in-place edit, so this was creating a backup
with the suffix `-r`. Providing an empty string skips the backup. The
`-r` for extended regex probably isn't needed in this case (given it
wasn't being used as a flag before), but to keep things consistent with
other `sed` usage across these script files, I'll keep it