Make locale configurable for links in GraphQL
We're en route to introducing another Content API-a-like behaviour to
the content items served from the GraphQL endpoint.

In Content Store, expanded links either match the locale of the root
Content Item or they default to "en" as a fallback. We'll shortly be
updating these arguments to pass in the root Edition's locale.