Get GraphQL links that match primary locale or "en"
In Content Store, expanded links either match the locale of the root
Content Item or they default to "en" as a fallback. This change updates
our GraphQL endpoint to make regular, forward-pointing links and reverse
links work the same way Content Store does.

I realise that it might not be immediately obvious what's happening in
this new code, especially because we don't use this sort of SQL a lot on
this project. Hopefully it'll be useful if I describe what we've
changed:

We've updated the WHEREs in the subqueries to find linked Editions
matching both the primary locale and the fallback ("en"). In most cases
the primary locale is "en", so we'd just be looking at 1 linked Edition.
There might be cases where there are no matches, too. So we should be
prepared for finding 0, 1 or 2 linked Editions per link, only differing
by locale.

Then we use `row_number()` to (basically) rank how well each candidate
Edition matches the primary locale and an additional WHERE filter on the
parent query picks out the best match of each (i.e. row_number 1)

If it's new to you, `row_number()` is a window function[^1]. Since it's
a member of the SELECT, its results will be included as a column on the
corresponding rows. (And they'll be called "row_number" because we
didn't alias it.)

The case statement produces a 0 or 1 for the `ORDER BY` to use for
sorting. Since ASC is the default sort order, the branch producing
a 0 will end up ranked highest.

[^1]: https://www.postgresql.org/docs/current/functions-window.html

---

Note that the special case of the MinistersIndexType's
person-role_appointment-role links (handled by PersonCurrentRolesSource)
has been left as is, with its hardcoded `locale: "en"`. As far as I can
tell, there are no non-"en" versions of the Ministers Index page and
therefore there's no reason to introduce the overhead to that specific
database query.

---

Special thanks to Richard for teaching me that you can use a case
statement in a window function's ORDER BY :O