Do not return the same document twice in reverse links
There are cases where the same links appear as both link set links and
edition links (e.g. document collections published by Whitehall [1]).

When sent to Content Store, the document is only included in the reverse
links array once.  However GraphQL is returning the link twice, meaning
the same link in repeated twice on the website.

Updating the logic of the source to only return the link once, even if
it appears as both a link set and edition link.

1: https://github.com/alphagov/whitehall/blob/03d5b51948b367d4484bd22ca78708eeb6dda017/app/presenters/publishing_api/document_collection_presenter.rb#L33-L48