Extract Link::PERMITTED_UNPUBLISHED_LINK_TYPES
When an Edition is unpublished, it shouldn't be served as a link...
unless the associated `link_type` is one of "children", "parent" or
"related_statistical_data_sets".

I've called this concept a "permitted unpublished link_type", since
unpublished Editions are permitted to be linked to only if they have
that link_type.

FWIW, I wasn't sure if it made more sense to replace those hard-coded
lists in the specs or to leave them.