Only request "en" locale reverse links
Officially, our GraphQL content only supports English language links
right now. That was already hard-coded in the other link dataloader
sources, but not this one.

We're in the process of making locales actually behave as expected, but
the next change in that feature will be more straightforward if this
WHERE clause is in place before we make it.