Ignore withdrawn timestamps in diff script
Content Store and GraphQL present this timestamp in different timezones
and with different string formatting.

E.g. on the page /government/news/zika-virus-data-update, Content Store
has the withdrawn time as "2019-06-25T14:36:51Z" while GraphQL has it as
"2019-06-25T15:36:51+01:00".

Because the GraphQL version is actually consistent with Content Store's
own usual timezone and time string formatting rules, we'll look into
fixing the issue in Content Store (possibly the underlying data) and for
the time being allow the difference.